Respect PsReadline's HistoryNoDuplicates setting
If a user has chosen to ignore duplicates in the Psreadline history,
apply this to the history shown by Invoke-FzfPsReadlineHandlerHistory as
well for consistency since it will usually used in place of the original
PsReadLine hisotroy search.
Note that PsReadLine ignores duplicates in code, not in the file it saves,
so we must do the same by filtering duplicates from the file.