Fix sort order when listing history without duplicates
Select-Object -Unique walks the input from top to bottom, which means
if we put in @(a, b, c, a) the output is @(a, b, c). This is unwanted
here because we're listing history and want the most recent commands to
turn up first, and in this case 'first' means at the bottom of the list;
in other words we want @(b, c, a) as output.
Fix this by reversing the history lines before removing duplicates.