Do not abort when failing to flush a stream
Exception during stream flush would abort the pipeline and fail to
properly process any result of selection when using Alt+c or Ctrl+t
functionality, for example.

This *will* happen on Linux when accepting the selection early, before
the earlier step finishes streaming input and if it is a system command
rather than powershell cmdlet.

Fixes #112.