Get-FileSystemCmd should honour dirOnly
Alt+C didn't really work with default commands unless fd was enabled,
because other default filesystem commands didn't know how to show
directories only and showed files instead.

In particular, FZF_DEFAULT_COMMAND cannot be used with dirOnly as you
never know what's there and how to adapt it. The script defaults now
have a dirOnly versions as well.

Additionally, find needs to at least follow symbolic links. Otherwise,
it cannot say if they should count as files or directories which leads
to being unable to switch to a directory that is a link, or showing
links to directories as file input which is not what you expect to
happen in such situations on Linux.

Fixes #114.