Do not abort when failing to flush a stream (#113)
* Do not abort when failing to flush a stream

Exception during stream flush would abort the pipeline and fail to
properly process any result of selection when using Alt+c or Ctrl+t
functionality, for example.

This *will* happen on Linux when accepting the selection early, before
the earlier step finishes streaming input and if it is a system command
rather than powershell cmdlet.

Fixes #112.

* Better handling of possibility of a broken pipe

Upon further examination, it seems that any time we write to
$utf8Stream, it may error out because the process may have exited
between last check of its status and stream action. So all operations on
it should guard against such scenario, however unlikely it may be.

This is implemented by ensuring $utf8Stream is in script scope so that
it can be accessed from a new scriptblock instead of repeating the same
few lines of code every time. I also removed assignment to $utf8Stream
from $cleanup because it wrote to the local scope not to the variable
in the parent scope, which is probably why almost every invocation of
$cleanup was also resetting $utf8Stream separately.