PSFzf.Base.ps1: enhance <Ctrl+t> functionality
Return a full path for the selected item on <Ctrl+t>. This solves the problem when the function is invoked with path, e.g. invoking it in current directory works fine, but when doing something like this: "ls C:\Path<Ctrl+t>" and selecting an item, will end up with "ls Sub-path". This change addresses the issue.