feat(preview-window): Allow for customization of preview window on tab completion
This commit is adding an environment variable to allow of customization
of the preview window during tab-completion.

If the user sets `$env:_PSFZF_FZF_TAB_PREVIEW_DEFAULT = "right"` then
the preview window during tab completion automatically shows to the
right instead of at the bottom.

This also keeps the Ctrl+/ binding in-tact by automatically switching
the options for the alternative views.